- wire up D2C control panel so that controls affect each other - change FourierSlider method to get/set double precision values - override getValue and setValue in FourierSlider subclasses instead of having specialized methods

+++ fourier/src/edu/colorado/phet/fourier/control/DiscreteControlPanel.java
@@ -621,7 +621,7 @@ public class DiscreteControlPanel extends FourierControlPanel implements ChangeL
         
         setWaitCursorEnabled( true );
         
-        int numberOfHarmonics = _numberOfHarmonicsSlider.getValue();
+        int numberOfHarmonics = (int)_numberOfHarmonicsSlider.getValue();
         
         // Update the Fourier series.
         _fourierSeries.setNumberOfHarmonics( numberOfHarmonics );
